+++ libs/common/src/ghostos_common/helpers/files.py
@@ -1,8 +1,9 @@
+from abc import abstractmethod
 import pathlib
-from typing import List, Tuple, Iterable, Dict, Union
+from typing import List, Tuple, Iterable, Dict, Union, Protocol, Any
 import fnmatch
 
-__all__ = ['list_dir', 'match_pathname', 'generate_directory_tree']
+__all__ = ['list_dir', 'is_pathname_ignored', 'generate_directory_tree', 'DescriptionsGetter']
 
 
 def list_dir(
@@ -47,10 +48,10 @@ def list_dir(
     # 3. 遍历目标目录
     for path in target_dir.iterdir():
         # 检查路径是否被忽略
-        if match_pathname(path.name, ignores):
+        if is_pathname_ignored(path.name, ignores, path.is_dir()):
             continue
 
-        if includes and not match_pathname(path.name, includes):
+        if includes and not is_pathname_ignored(path.name, includes):
             continue
 
         # 如果是文件且 files=True，则返回
@@ -66,7 +67,7 @@ def list_dir(
                 )
 
 
-def match_pathname(path: Union[pathlib.Path, str], pattern: Iterable[str]) -> bool:
+def is_pathname_ignored(path: Union[pathlib.Path, str], pattern: Iterable[str], is_dir: bool) -> bool:
     if not pattern:
         return False
     if isinstance(path, pathlib.Path):
@@ -74,15 +75,28 @@ def match_pathname(path: Union[pathlib.Path, str], pattern: Iterable[str]) -> bo
     else:
         name = str(path)
     for pattern in pattern:
+        matched = True
+        if pattern.startswith('!'):
+            matched = False
+            pattern = pattern[1:]
+        if is_dir and pattern.endswith('/'):
+            pattern = pattern[:-1]
         if fnmatch.fnmatch(name, pattern):
-            return True
+            return matched
     return False
+
+
+class DescriptionsGetter(Protocol):
+
+    @abstractmethod
+    def get(self, path: pathlib.Path, default: Union[str, None] = None) -> Union[str, None]:
+        pass
 
 
 def generate_directory_tree(
         current: Union[pathlib.Path, str],
         recursion: int = -1,
-        descriptions: Dict[str, str] = None,
+        descriptions: Union[Dict[str, str], DescriptionsGetter, None] = None,
         *,
         prefix: str = "",
         ignores: List[str] = None,
@@ -118,17 +132,22 @@ def generate_directory_tree(
         current_indent = indent * current_depth
 
         desc = ""
-        if descriptions:
+        if descriptions is not None:
             relative = path.relative_to(current)
             relative_path = str(relative)
-            if relative_path in descriptions:
-                got = descriptions.get(relative_path, "")
-                got.strip()
+            desc = ""
+            got = ""
+            if isinstance(descriptions, dict):
+                if relative_path in descriptions:
+                    got = descriptions.get(relative_path, "")
+            else:
+                got = descriptions.get(path, "")
+            got.strip()
+            if got:
                 got = got.replace("\n", " ")
                 if len(got) > 150:
                     got = got[:150] + "..."
-                if got:
-                    desc = f" : `{got}`"
+                desc = f" : `{got}`"
 
         if path.is_dir():
             tree.append(f"{current_indent}📁 {path.name}{desc}")