+++ libs/ghostos/ghostos/ghosts/project_manager/__init__.py
@@ -1,0 +1,3 @@
+from ghostos.ghosts.project_manager.project_manager_ghost import ProjectManagerProvider, ProjectManagerGhost
+
+__ghost__ = ProjectManagerGhost.new(None)