+++ libs/ghostos/ghostos/libraries/project/__init__.py
@@ -3,6 +3,7 @@ from ghostos.libraries.project.abcd import (
     Directory,
     File,
     PyDevCtx,
+    ProjectExports,
 )
 from ghostos.libraries.project import abcd as contracts
 from ghostos.libraries.project.project_impl import ProjectManagerProvider