+++ libs/common/RELEASES.md
@@ -3,6 +3,19 @@
 move `ghostos.prompter` to `ghostos_common.prompter`.
 provide prompt object model for `ghosots_moss`.
 
+## 0.3.4
+
+`generate_directory_tree` accept `descriptionGetter`
+
+## 0.3.3 
+
+fix: fix `list_dir` can not ignore directory in `.gitignore`.
+
+## 0.3.2
+
+* add `get_module_fullname_from_path` method. 
+* fix `list_dir` can't ignore dir
+
 ## 0.3.1
 
 * fix `generate_module_and_attr_name` can not generate from object decorator.